requirements.txt
arrow==0.15.8
asgiref==3.2.10
Babel==2.8.0
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
Django==3.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-extensions==2.2.9
django-phonenumber-field==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
phonenumbers==8.12.7
Pillow==7.1.1
poyo==0.5.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
raven==6.9.0
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
